refactor: extract server-admin.js from server.js (PR-16)
Move all admin API endpoints into dedicated server-admin module:
user CRUD, permissions, PIN reset, Linux user mapping, invites,
SMTP config/test, and project access control (visibility, owner,
allowed users).

server.js delegates admin routes via admin.handleRequest() and
passes dependencies (users, smtp, auth helpers) through context.

server.js: 3086 -> 2428 lines. server-admin.js: 712 lines.
